backend/requirements.txt
# FastAPI and ASGI server
fastapi>=0.95.0,<0.96.0
uvicorn>=0.22.0,<0.23.0
pydantic>=1.10.7,<1.11.0
email-validator>=2.0.0,<2.1.0

# Database
sqlalchemy>=2.0.9,<2.1.0
alembic>=1.10.3,<1.11.0
mysql-connector-python>=8.0.32,<8.1.0

# Security
python-jose>=3.3.0,<3.4.0
passlib>=1.7.4,<1.8.0
bcrypt>=4.0.1,<4.1.0
pyotp>=2.8.0,<2.9.0
python-multipart>=0.0.6,<0.1.0

# Redis (for rate limiting and caching)
redis>=4.5.4,<4.6.0

# Email
fastapi-mail>=1.2.8,<1.3.0

# Testing
pytest>=7.3.1,<7.4.0
pytest-cov>=4.1.0,<4.2.0
httpx>=0.24.0,<0.25.0

# Utilities
python-dotenv>=1.0.0,<1.1.0
tenacity>=8.2.2,<8.3.0